Multiple changes for siamese, see below:
- Now plots the loss from fitting
- Limits the number of training pairs generated. Necessary
because it grows combinatorially, so you may run out of memory if
you don't limit it
- Added different margins for Contrastive Loss
- Added split_dataset.py script to split a csv into train and test
sets that have the same distribution of experiments.

+++ data_container.py
@@ -49,11 +49,11 @@ class DataContainer(object):
         return self.dataframe.columns.values[2:]
 
     def get_all_dataset_IDs(self):
-        return self.dataframe.loc[:, 'Dataset'].values
+        return self.dataframe.loc[:, 'Dataset'].values.astype(int)
 
     def get_labeled_dataset_IDs(self):
         labeled_data = self.dataframe.loc[lambda df: df.Label != 'None', :]
-        return labeled_data.loc[:, 'Dataset'].values
+        return labeled_data.loc[:, 'Dataset'].values.astype(int)
 
     def get_all_sample_IDs(self):
         #return self.dataframe.loc[:, 'Sample'].values
